refactor: status endpoint package renamed
