+++ webhook-service/src/main/scala/io/renku/webhookservice/eventstatus/HookEventEndpoint.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.renku.webhookservice.eventprocessing
+package io.renku.webhookservice.eventstatus
 
 import cats.data.NonEmptyList
 import cats.effect._
